refactor: split skill writer consolidation
