+++ tests/agent_suite/managed_skills_test.rs
@@ -2,10 +2,9 @@ use tracedecay::automation::managed_skills::{
     approve_managed_skill, archive_managed_skill, create_managed_skill_draft,
     disable_managed_skill, discard_pending_managed_skill_update, list_managed_skills,
     load_managed_skill, managed_skill_dir, restore_managed_skill, save_managed_skill,
-    set_managed_skill_state, stage_managed_skill_archive, stage_managed_skill_update,
-    update_managed_skill, ManagedSkillDraft, ManagedSkillProvenance, ManagedSkillSource,
-    ManagedSkillState, ManagedSkillUpdate, ManagedSupportFile, SkillInstallTarget,
-    MAX_MANAGED_SKILL_BODY_BYTES,
+    set_managed_skill_state, stage_managed_skill_update, update_managed_skill, ManagedSkillDraft,
+    ManagedSkillProvenance, ManagedSkillSource, ManagedSkillState, ManagedSkillUpdate,
+    ManagedSupportFile, SkillInstallTarget, MAX_MANAGED_SKILL_BODY_BYTES,
 };
 use tracedecay::automation::skill_usage::{
     ingest_analytics_events, load_skill_usage_records, record_skill_usage,
@@ -836,142 +835,6 @@ async fn staged_managed_skill_update_preserves_active_revision_until_approval() 
         "Stage safer repository hygiene guidance."
     );
     assert_eq!(reloaded.metadata.checksum, second_base);
-}
-
-#[tokio::test]
-async fn staged_managed_skill_archive_keeps_content_until_approval() {
-    let temp = tempfile::tempdir().unwrap();
-    let profile_root = temp.path().join("profile");
-    create_managed_skill_draft(&profile_root, draft())
-        .await
-        .unwrap();
-    let active = approve_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    let base_checksum = active.metadata.checksum.clone();
-    let skill_dir = managed_skill_dir(&profile_root, "repo-hygiene").unwrap();
-
-    let staged = stage_managed_skill_archive(
-        &profile_root,
-        "repo-hygiene",
-        &base_checksum,
-        Some("overlaps with newer guidance".to_string()),
-    )
-    .await
-    .unwrap();
-    assert_eq!(staged.metadata.state, ManagedSkillState::PendingApproval);
-
-    // Staging must not touch the active revision.
-    let with_pending = load_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    assert_eq!(with_pending.metadata.state, ManagedSkillState::Active);
-    assert_eq!(with_pending.metadata.checksum, base_checksum);
-    let pending = with_pending.pending_update.as_ref().unwrap();
-    assert_eq!(pending.resulting_state, Some(ManagedSkillState::Archived));
-    assert_eq!(
-        pending.staged_reason.as_deref(),
-        Some("overlaps with newer guidance")
-    );
-
-    // Approval archives without deleting any content.
-    let approved = approve_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    assert_eq!(approved.metadata.state, ManagedSkillState::Archived);
-    assert_eq!(approved.body_markdown, active.body_markdown);
-    assert!(approved.pending_update.is_none());
-    assert!(skill_dir.join("SKILL.md").is_file());
-    assert!(skill_dir.join("references/checklist.md").is_file());
-    let reloaded = load_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    assert_eq!(reloaded.metadata.state, ManagedSkillState::Archived);
-    assert_eq!(reloaded.body_markdown, active.body_markdown);
-}
-
-#[tokio::test]
-async fn staged_managed_skill_archive_can_be_discarded() {
-    let temp = tempfile::tempdir().unwrap();
-    let profile_root = temp.path().join("profile");
-    create_managed_skill_draft(&profile_root, draft())
-        .await
-        .unwrap();
-    let active = approve_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    stage_managed_skill_archive(
-        &profile_root,
-        "repo-hygiene",
-        &active.metadata.checksum,
-        None,
-    )
-    .await
-    .unwrap();
-
-    let discarded = discard_pending_managed_skill_update(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    assert!(discarded.pending_update.is_none());
-    let reloaded = load_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    assert_eq!(reloaded.metadata.state, ManagedSkillState::Active);
-    assert_eq!(reloaded.metadata.checksum, active.metadata.checksum);
-}
-
-#[tokio::test]
-async fn staged_managed_skill_archive_rejects_pinned_stale_and_duplicates() {
-    let temp = tempfile::tempdir().unwrap();
-    let profile_root = temp.path().join("profile");
-    create_managed_skill_draft(&profile_root, draft())
-        .await
-        .unwrap();
-    let mut active = approve_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    let base_checksum = active.metadata.checksum.clone();
-
-    let err = stage_managed_skill_archive(&profile_root, "repo-hygiene", "sha256:stale", None)
-        .await
-        .unwrap_err();
-    assert!(err.to_string().contains("is stale"));
-
-    active.set_pinned(true);
-    save_managed_skill(&profile_root, &active).await.unwrap();
-    let err = stage_managed_skill_archive(&profile_root, "repo-hygiene", &base_checksum, None)
-        .await
-        .unwrap_err();
-    assert!(err
-        .to_string()
-        .contains("pinned and exempt from staged archive"));
-
-    active.set_pinned(false);
-    save_managed_skill(&profile_root, &active).await.unwrap();
-    stage_managed_skill_archive(&profile_root, "repo-hygiene", &base_checksum, None)
-        .await
-        .unwrap();
-    let err = stage_managed_skill_archive(&profile_root, "repo-hygiene", &base_checksum, None)
-        .await
-        .unwrap_err();
-    assert!(err.to_string().contains("already has a pending update"));
-
-    // Already-archived skills cannot be staged for archive again.
-    discard_pending_managed_skill_update(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    let archived = archive_managed_skill(&profile_root, "repo-hygiene")
-        .await
-        .unwrap();
-    let err = stage_managed_skill_archive(
-        &profile_root,
-        "repo-hygiene",
-        &archived.metadata.checksum,
-        None,
-    )
-    .await
-    .unwrap_err();
-    assert!(err.to_string().contains("is already archived"));
 }
 
 #[tokio::test]
